Add explicit options for `json.dumps()` #minor
Set explicit options for `json.dumps()`. Biggest change is to match the
separators with ones used in ECMAScript for better overall
interoperability. The character set is also no longer transcoded to
ascii.

The API remains the same, but hashes prior to v0.12.0 will not be
comparable. Added also a notice to README.md to emphasize this aspect.