Fix empty `list` and empty `dict` handling
Empty lists and empty dicts were not previously considered to be values.
This caused the entire path to an empty list or dict to be ignored.

JSON objects with or without empty list or dict values could cause jfpv1
hash collisions with similar objects up to version `0.12.2`. Any release
beyond `v0.12.z` should be considered backward-incompatible for cases
where empty lists or dicts are present.

See the following issue for more details:
https://github.com/cobaltine/json-fingerprint/issues/33